Fix for mask not applying to output IDW raster
For unknown reason arcpy.env.mask was not being honored. The workaround is to clip the data to the mask layer, which also requires an extent minxy,maxxy. Removed redundant import. Added triple dots in print statement to indicate the script is processing. Using env cellsize instead of one variable for IDW function.